requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
colorama==0.4.4
flasgger==0.9.5
Flask==2.0.2
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
mistune==0.8.4
psycopg2==2.9.2
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
Werkzeug==2.0.2
